notification-service/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.1
billiard==4.1.0
blinker==1.6.2
cachelib==0.9.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.3.0
eventlet==0.33.3
exceptiongroup==1.1.1
Flask==2.3.2
Flask-Celery==2.4.3
Flask-Mail==0.9.1
Flask-Script==2.0.6
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
kombu==5.3.1
Mako==1.2.4
MarkupSafe==2.1.3
mistune==3.0.1
packaging==23.1
passlib==1.7.4
pip-autoremove==0.10.0
prompt-toolkit==3.0.38
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-engineio==4.4.1
pytz==2023.3
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.13
uvicorn==0.22.0
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
zipp==3.15.0
